+++ example/lib/main.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:schema_widget/schema_widget.dart';
 
@@ -8,7 +10,6 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return SchemaWidget.build(context, {
@@ -25,15 +26,6 @@ class MyApp extends StatelessWidget {
 class MyHomePage extends StatefulWidget {
   MyHomePage({Key key, this.title}) : super(key: key);
 
-  // This widget is the home page of your application. It is stateful, meaning
-  // that it has a State object (defined below) that contains fields that affect
-  // how it looks.
-
-  // This class is the configuration for the state. It holds the values (in this
-  // case the title) provided by the parent (in this case the App widget) and
-  // used by the build method of the State. Fields in a Widget subclass are
-  // always marked "final".
-
   final String title;
 
   @override
@@ -43,20 +35,40 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   int _counter = 0;
 
+  final StreamController<int> _streamControllerInt = StreamController<int>();
+
+  _MyHomePageState() {
+    SchemaWidget.registerLogic("streamInt", _streamControllerInt.stream);
+    SchemaWidget.registerLogic("buildTextCounter", _buildTextCounter);
+  }
+
   void _incrementCounter() {
-    setState(() {
-      // This call to setState tells the Flutter framework that something has
-      // changed in this State, which causes it to rerun the build method below
-      // so that the display can reflect the updated values. If we changed
-      // _counter without calling setState(), then the build method would not be
-      // called again, and so nothing would appear to happen.
-      _counter++;
+    _streamControllerInt.add(++_counter);
+  }
+
+  Widget _buildTextCounter(
+      BuildContext buildContext, AsyncSnapshot<dynamic> snapshot) {
+    var textStyle = Theme.of(buildContext).textTheme.display1;
+
+    return SchemaWidget.build(buildContext, {
+      "type": "Text",
+      "data": '${snapshot.data}',
+      "style": {
+        "type": "TextStyle",
+        "color": "#${intToHex(textStyle.color.value, 6)}",
+        "debugLabel": textStyle.debugLabel,
+        "decoration": textDecorationToString(textStyle.decoration),
+        "fontFamily": textStyle.fontFamily,
+        "fontSize": textStyle.fontSize,
+        "fontWeight": textStyle.fontWeight.toString().substring(11),
+        "fontStyle":
+            textStyle.fontStyle == FontStyle.italic ? 'italic' : 'normal',
+      },
     });
   }
 
   @override
   Widget build(BuildContext context) {
-    var textStyle = Theme.of(context).textTheme.display1;
     var iconData = Icons.add;
 
     return SchemaWidget.build(context, {
@@ -79,20 +91,10 @@ class _MyHomePageState extends State<MyHomePage> {
               "data": 'You have pushed the button this many times:',
             },
             {
-              "type": "Text",
-              "data": '$_counter',
-              "style": {
-                "type": "TextStyle",
-                "color": "#${intToHex(textStyle.color.value, 6)}",
-                "debugLabel": textStyle.debugLabel,
-                "decoration": textDecorationToString(textStyle.decoration),
-                "fontFamily": textStyle.fontFamily,
-                "fontSize": textStyle.fontSize,
-                "fontWeight": textStyle.fontWeight.toString().substring(11),
-                "fontStyle": textStyle.fontStyle == FontStyle.italic
-                    ? 'italic'
-                    : 'normal',
-              },
+              "type": "StreamBuilder",
+              "stream": "streamInt",
+              "initialData": _counter,
+              "builder": "buildTextCounter",
             },
           ],
         },